Split teams.js into routes/teams/ sub-routers (Chunk 9)
Pure move, no behaviour change. teams.js (1,267 lines) split into
routes/teams/ (list.js, crud.js, members.js, events.js, ledger.js,
index.js) following the groups/ sub-router precedent. index.js owns the
shared requireAuth + requireFeature('teams') middleware, then mounts the
sub-routers with the list router (literal /download) before the CRUD
router (/:id) to preserve Express match order. app.js import updated.
