+++ backend/src/app.js
@@ -22,7 +22,7 @@ import memberRoutes from './routes/members/index.js';
 import facultyRoutes from './routes/faculties.js';
 import venueRoutes from './routes/venues.js';
 import groupRoutes from './routes/groups/index.js';
-import teamRoutes from './routes/teams.js';
+import teamRoutes from './routes/teams/index.js';
 import settingsRoutes from './routes/settings.js';
 import financeRoutes from './routes/finance/index.js';
 import pollRoutes from './routes/polls.js';
